+++ packages/plugins/pdf/test/goldens/pdf_export_dialog_golden_test.dart
@@ -4,6 +4,7 @@ import 'package:flutter/material.dart';
 import 'package:flutter_test/flutter_test.dart';
 import 'package:superdeck/superdeck.dart';
 import 'package:superdeck_core/superdeck_core.dart';
+import 'package:superdeck_pdf/superdeck_pdf.dart';
 import 'package:superdeck_pdf/src/pdf_export_screen.dart';
 
 import '../helpers/test_helpers.dart';
@@ -49,10 +50,12 @@ void main() {
                     ),
                   ),
                 ],
-                pdfSaver: (_) {
-                  saverCalled = true;
-                  return releaseSave.future;
-                },
+                options: PdfExportOptions(
+                  pdfSaver: (_) {
+                    saverCalled = true;
+                    return releaseSave.future;
+                  },
+                ),
               ),
             ),
           ),