+++ packages/superdeck/test/public_api_test.dart
@@ -13,7 +13,9 @@ void main() {
       expect(ThumbnailService, isNotNull);
       expect(superDeckSlideSize, isNotNull);
       expect(superDeckAspectRatio, isNotNull);
+      expect(DeckPlugin, isNotNull);
       expect(DeckAction, isNotNull);
+      expect(DeckRuntimePlugin, isNotNull);
     });
   });
 }